[dev] Change abstraction structure for bbdl_base_domain
